+++ src/components/forms/ColorPalette/ColorPalette.story.tsx
@@ -1,4 +1,5 @@
-import { ColorPalette } from "./ColorPalette";
+import { ColorPalette } from ".";
+
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof ColorPalette> = {
